Feat: Generate the right posts by tag
